requirements.txt
Django==5.2
Pillow==11.2.1
asgiref==3.8.1
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
web3==6.2.0
eth-utils==2.1.0
razorpay==1.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9 